re-structure utilities files in python client
