+++ cytomine/models/_utilities/__init__.py
@@ -1,0 +1,5 @@
+
+
+from .download import generic_download, makedirs, is_false
+from .dump import generic_image_dump, DumpError
+from .pattern_matching import resolve_pattern, is_iterable
