+++ cytomine/models/_utilities/download.py
