+++ cytomine/models/_utilities/pattern_matching.py
