--- cytomine/utilities/logging.py
@@ -1,34 +1,0 @@
-# -*- coding: utf-8 -*-
-
-# * Copyright (c) 2009-2018. Authors: see NOTICE file.
-# *
-# * Licensed under the Apache License, Version 2.0 (the "License");
-# * you may not use this file except in compliance with the License.
-# * You may obtain a copy of the License at
-# *
-# *      http://www.apache.org/licenses/LICENSE-2.0
-# *
-# * Unless required by applicable law or agreed to in writing, software
-# * distributed under the License is distributed on an "AS IS" BASIS,
-# * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# * See the License for the specific language governing permissions and
-# * limitations under the License.
-
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
-from __future__ import unicode_literals
-
-import sys
-
-__author__ = "Rubens Ulysse <urubens@uliege.be>"
-__copyright__ = "Copyright 2010-2018 University of Liège, Belgium, http://www.cytomine.be/"
-
-from logging import StreamHandler, Formatter
-
-
-class StdoutHandler(StreamHandler):
-    def __init__(self):
-        super(StreamHandler, self).__init__()
-        self.stream = sys.stdout
-        self.setFormatter(Formatter("[%(asctime)s][%(levelname)s] %(message)s"))
